refactor: move http submission to dedicate namespace
Also:
- remove Newtonsoft dependencies on the models
- unify http summition request and base request
